non-LTR/mgescan/requirements.txt
docopt==0.6.1
bcbio-gff
biopython
